Returning all track information and displaying as expected
