--- spotify-timer-frontend/src/app/current-track/current-track.component.css
@@ -1,23 +1,0 @@
-.glass-container {
-    width: 100%; /* Set a fixed width */
-    height: 200px; /* Set a fixed height */
-    background-color: rgba(0, 0, 0, 0.6); /* Semi-transparent black */
-    border-radius: 15px; /* Rounded corners */
-    backdrop-filter: blur(10px); /* Frosted glass effect */
-    -webkit-backdrop-filter: blur(10px); /* For Safari support */
-    border: 1px solid rgba(255, 255, 255, 0.2); /* Optional: Light border for contrast */
-    box-shadow: 0 4px 20px rgba(0, 0, 0, 0.5); /* Shadow for depth */
-    margin: 20px auto; /* Center the container */
-    display: flex; /* Flexbox for inner content alignment */
-    align-items: center; /* Center vertically */
-    justify-content: center; /* Center horizontally */
-    color: white; /* Text color */
-    font-family: 'Gotham', sans-serif; /* Font family */
-}
-
-.track-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 100%; /* Ensure it fits in the parent container */
-}
