--- spotify-timer-frontend/src/app/current-track/current-track.component.html
@@ -1,23 +1,0 @@
-<div class="glass-container">
-    <div class="track-container"> 
-        <h1>Song Name</h1>
-        <h2>Album</h2>
-        <h3>Artist</h3>
-    </div>    
-</div>
-
-<!-- current-track.component.html -->
-
-<div class="current-track-container">
-    <h2 *ngIf="currentTrack">Now Playing: {{ currentTrack?.name }}</h2>
-    <p *ngIf="currentTrack">Artist(s): {{ currentTrackArtists }}</p>
-    
-  
-    <div class="button-container">
-      <button (click)="isPlaying ? pausePlayback() : playCurrentTrack()">
-        {{ isPlaying ? 'Pause' : 'Play' }}
-      </button>
-      <button (click)="onNext()">Next Track</button>
-    </div>
-  </div>
-
